+++ src/main/java/kitchenpos/menu/dto/request/MenuProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto.request;
 
 
 import javax.validation.constraints.Min;
