+++ src/main/java/kitchenpos/menu/dto/response/MenuResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.dto.response;
+package kitchenpos.menu.dto.response;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public class MenuResponse {
