+++ src/main/java/kitchenpos/menu/repository/MenuProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
